requirements.txt
Flask
mysql-connector-python
requests
jsonify
Flask-Cors
pymysql 

